CardBuilder & Card Data
Saves all possible cards in a single file to load from, use CardBuilder to get cards from the list.
This simplifies the saved structure of decks that I'll have to write soon, it also simplifies everything else a bit.
The CardBuilder loads the cards on ready, might need a refresh button/system later on just in case people want to customize.
It's also not future-proof, might need to add some verification system or other later.